test: move ami unit tests to its own package
