+++ pkg/handlers/aws/mutation/ami/inject_suite_test.go
@@ -1,0 +1,13 @@
+package ami
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestAMIPatch(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "AMI patches for ControlPlane and Workers suite")
+}
