+++ pkg/handlers/aws/mutation/ami/inject_worker_test.go
@@ -1,12 +1,12 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package tests
+package ami
 
 import (
-	"testing"
-
+	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 
@@ -14,99 +14,23 @@ import (
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
+	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/workerconfig"
 )
 
-func TestControlPlaneGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
+var _ = Describe("Generate AMI patches for Worker", func() {
+	patchGenerator := func() mutation.GeneratePatches {
+		return mutation.NewMetaGeneratePatchesHandler("", NewWorkerPatch()).(mutation.GeneratePatches)
+	}
 
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
-			Name: "AMI set for control plane",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					v1alpha1.AMISpec{ID: "ami-controlplane"},
-					variablePath...,
-				),
-			},
-			RequestItem: request.NewCPAWSMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{
-				{
-					Operation:    "add",
-					Path:         "/spec/template/spec/ami/id",
-					ValueMatcher: gomega.Equal("ami-controlplane"),
-				},
-			},
-		},
-		capitest.PatchTestDef{
-			Name: "AMI lookup format set for control plane",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					v1alpha1.AMISpec{
-						Lookup: &v1alpha1.AMILookup{
-							Format: "test-{{.kubernetesVersion}}-format",
-							Org:    "12345",
-							BaseOS: "testOS",
-						},
-					},
-					variablePath...,
-				),
-			},
-			RequestItem: request.NewCPAWSMachineTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{
-				{
-					Operation:    "add",
-					Path:         "/spec/template/spec/imageLookupFormat",
-					ValueMatcher: gomega.Equal("test-{{.kubernetesVersion}}-format"),
-				},
-				{
-					Operation:    "add",
-					Path:         "/spec/template/spec/imageLookupOrg",
-					ValueMatcher: gomega.Equal("12345"),
-				},
-				{
-					Operation:    "add",
-					Path:         "/spec/template/spec/imageLookupBaseOS",
-					ValueMatcher: gomega.Equal("testOS"),
-				},
-			},
-			UnexpectedPatchMatchers: []capitest.JSONPatchMatcher{
-				{
-					Operation:    "add",
-					Path:         "/spec/template/spec/ami/id",
-					ValueMatcher: gomega.Equal(""),
-				},
-			},
-		},
-	)
-}
-
-func TestWorkerGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
-
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
+	testDefs := []capitest.PatchTestDef{
+		{
 			Name: "AMI set for workers",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
-					variableName,
+					workerconfig.MetaVariableName,
 					v1alpha1.AMISpec{ID: "ami-controlplane"},
-					variablePath...,
+					v1alpha1.AWSVariableName,
+					VariableName,
 				),
 				capitest.VariableWithValue(
 					"builtin",
@@ -124,11 +48,11 @@ func TestWorkerGeneratePatches(
 				},
 			},
 		},
-		capitest.PatchTestDef{
+		{
 			Name: "AMI lookup format set for worker",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
-					variableName,
+					workerconfig.MetaVariableName,
 					v1alpha1.AMISpec{
 						Lookup: &v1alpha1.AMILookup{
 							Format: "test-{{.kubernetesVersion}}-format",
@@ -136,8 +60,8 @@ func TestWorkerGeneratePatches(
 							BaseOS: "testOS",
 						},
 					},
-
-					variablePath...,
+					v1alpha1.AWSVariableName,
+					VariableName,
 				),
 				capitest.VariableWithValue(
 					"builtin",
@@ -172,5 +96,17 @@ func TestWorkerGeneratePatches(
 				},
 			},
 		},
-	)
-}
+	}
+
+	// create test node for each case
+	for testIdx := range testDefs {
+		tt := testDefs[testIdx]
+		It(tt.Name, func() {
+			capitest.AssertGeneratePatches(
+				GinkgoT(),
+				patchGenerator,
+				&tt,
+			)
+		})
+	}
+})
